refactor(pi-terminal-mux): U2 每后端一个文件 + surface.ts 注册式派发
- 创建 backends/types.ts：BackendOps 内部接口（不对外导出）
- 创建 backends/muxy/cmux/tmux/zellij/wezterm.ts：各后端实现与 const ops
- 移动 herdr/otty → backends/，各补 BackendOps 适配器，原位置留 barrel
- 创建 surface.ts：统一 surface API + Record<MuxBackend, BackendOps> 全键注册表
- 重写 mux.ts 为瘦 barrel（84 行，仅 re-export）
- index.ts 内部 import 路径更新，公开导出符号集合不变
- 26 单测全绿，81 导出符号与基线 diff 无差异
- grep 验收通过（mux.ts 中无 createZellijSurfaceUnlocked/readCmux 命中）
